+++ relevanceai/api/batch/batch_insert.py
@@ -208,8 +208,9 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
         if "_id" not in chunk.columns:
             raise MissingFieldError("Need _id as a column")
 
-        # Convert vector fields
-        vector_columns = [i for i in chunk.columns if i.endswith("_vector_")]
+        # add fix for when lists are read in as strings
+        EXCEPTION_COLUMNS = ("_vector_", "_chunk_")
+        vector_columns = [i for i in chunk.columns if i.endswith(EXCEPTION_COLUMNS)]
         for i in vector_columns:
             chunk[i] = chunk[i].apply(literal_eval)
 
@@ -303,9 +304,9 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
 
     def pull_update_push(
         self,
-        original_collection: str,
+        dataset_id: str,
         update_function,
-        updated_collection: str = None,
+        updated_dataset_id: str = None,
         log_file: str = None,
         updating_args: dict = {},
         retrieve_chunk_size: int = 100,
@@ -321,14 +322,12 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
 
         Parameters
         ----------
-        original_collection : string
+        dataset_id: string
             The dataset_id of the collection where your original documents are
-        logging_collection: string
-            The dataset_id of the collection which logs which documents have been updated. If 'None', then one will be created for you.
-        updated_collection: string
-            The dataset_id of the collection where your updated documents are uploaded into. If 'None', then your original collection will be updated.
         update_function: function
             A function created by you that converts documents in your original collection into the updated documents. The function must contain a field which takes in a list of documents from the original collection. The output of the function must be a list of updated documents.
+        updated_dataset_id: string
+            The dataset_id of the collection where your updated documents are uploaded into. If 'None', then your original collection will be updated.
         updating_args: dict
             Additional arguments to your update_function, if they exist. They must be in the format of {'Argument': Value}
         retrieve_chunk_size: int
@@ -348,7 +347,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
         # Check if a logging_collection has been supplied
         if log_file is None:
             log_file = (
-                original_collection
+                dataset_id
                 + "_"
                 + str(datetime.now().strftime("%d-%m-%Y-%H-%M-%S"))
                 + "_pull_update_push"
@@ -363,7 +362,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
 
         # Trust the process
         # Get document lengths to calculate iterations
-        original_length = self.get_number_of_documents(original_collection, filters)
+        original_length = self.get_number_of_documents(dataset_id, filters)
 
         # get the remaining number in case things break
         remaining_length = original_length - PULL_UPDATE_PUSH_LOGGER.count_ids_in_fn()
@@ -386,7 +385,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
         ):
 
             orig_json = self.datasets.documents.get_where(
-                original_collection,
+                dataset_id,
                 filters=retrieve_filters,
                 page_size=retrieve_chunk_size,
                 select_fields=select_fields,
@@ -404,9 +403,9 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
             updated_documents = [i["_id"] for i in documents]
 
             # Upload documents
-            if updated_collection is None:
+            if updated_dataset_id is None:
                 insert_json = self.update_documents(
-                    dataset_id=original_collection,
+                    dataset_id=dataset_id,
                     docs=updated_data,
                     max_workers=max_workers,
                     show_progress_bar=False,
@@ -414,7 +413,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
                 )
             else:
                 insert_json = self.insert_documents(
-                    dataset_id=updated_collection,
+                    dataset_id=updated_dataset_id,
                     docs=updated_data,
                     max_workers=max_workers,
                     show_progress_bar=False,
@@ -440,10 +439,10 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
 
     def pull_update_push_to_cloud(
         self,
-        original_collection: str,
+        dataset_id: str,
         update_function,
-        updated_collection: str = None,
-        logging_collection: str = None,
+        updated_dataset_id: str = None,
+        logging_dataset_id: str = None,
         updating_args: dict = {},
         retrieve_chunk_size: int = 100,
         retrieve_chunk_size_failure_retry_multiplier: float = 0.5,
@@ -461,11 +460,11 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
 
         Parameters
         ----------
-        original_collection : string
+        original_dataset_id: string
             The dataset_id of the collection where your original documents are
-        logging_collection: string
+        logging_dataset_id: string
             The dataset_id of the collection which logs which documents have been updated. If 'None', then one will be created for you.
-        updated_collection: string
+        updated_dataset_id: string
             The dataset_id of the collection where your updated documents are uploaded into. If 'None', then your original collection will be updated.
         update_function: function
             A function created by you that converts documents in your original collection into the updated documents. The function must contain a field which takes in a list of documents from the original collection. The output of the function must be a list of updated documents.
@@ -483,9 +482,9 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
             Whether to automatically convert documents to json encodable format
         """
         # Check if a logging_collection has been supplied
-        if logging_collection is None:
-            logging_collection = (
-                original_collection
+        if logging_dataset_id is None:
+            logging_dataset_id = (
+                dataset_id
                 + "_"
                 + str(datetime.now().strftime("%d-%m-%Y-%H-%M-%S"))
                 + "_pull_update_push"
@@ -493,9 +492,9 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
 
         # Check collections and create completed list if needed
         collection_list = self.datasets.list()
-        if logging_collection not in collection_list["datasets"]:
+        if logging_dataset_id not in collection_list["datasets"]:
             self.logger.info("Creating a logging collection for you.")
-            self.logger.info(self.datasets.create(logging_collection))
+            self.logger.info(self.datasets.create(logging_dataset_id))
 
         # Track failed documents
         failed_documents: List[Dict] = []
@@ -504,8 +503,8 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
         for _ in range(number_of_retrieve_retries):
 
             # Get document lengths to calculate iterations
-            original_length = self.get_number_of_documents(original_collection, filters)
-            completed_length = self.get_number_of_documents(logging_collection)
+            original_length = self.get_number_of_documents(dataset_id, filters)
+            completed_length = self.get_number_of_documents(logging_dataset_id)
             remaining_length = original_length - completed_length
             iterations_required = math.ceil(remaining_length / retrieve_chunk_size)
 
@@ -518,7 +517,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
                 self.logger.success(f"Pull, Update, Push is complete!")
                 return {
                     "failed_documents": failed_documents,
-                    "logging_collection": logging_collection,
+                    "logging_collection": logging_dataset_id,
                 }
 
             for _ in progress_bar(
@@ -526,7 +525,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
             ):
 
                 # Get completed documents
-                log_json = self.get_all_documents(logging_collection)
+                log_json = self.get_all_documents(logging_dataset_id)
                 completed_documents_list = [i["_id"] for i in log_json]
 
                 # Get incomplete documents from raw collection
@@ -540,7 +539,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
                 ]
 
                 orig_json = self.datasets.documents.get_where(
-                    original_collection,
+                    dataset_id,
                     filters=retrieve_filters,
                     page_size=retrieve_chunk_size,
                     select_fields=select_fields,
@@ -560,9 +559,9 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
                 self.logger.debug(f"{len(updated_data)}")
 
                 # Upload documents
-                if updated_collection is None:
+                if updated_dataset_id is None:
                     insert_json = self.update_documents(
-                        dataset_id=original_collection,
+                        dataset_id=dataset_id,
                         docs=updated_data,
                         max_workers=max_workers,
                         show_progress_bar=False,
@@ -570,7 +569,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
                     )
                 else:
                     insert_json = self.insert_documents(
-                        dataset_id=updated_collection,
+                        dataset_id=updated_dataset_id,
                         docs=updated_data,
                         max_workers=max_workers,
                         show_progress_bar=False,
@@ -587,7 +586,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
                 upload_documents = [{"_id": i} for i in success_documents]
 
                 self.insert_documents(
-                    logging_collection,
+                    logging_dataset_id,
                     upload_documents,
                     max_workers=max_workers,
                 )
@@ -610,13 +609,13 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
                     )
                     return {
                         "failed_documents": failed_documents,
-                        "logging_collection": logging_collection,
+                        "logging_collection": logging_dataset_id,
                     }
 
         self.logger.success(f"Pull, Update, Push is complete!")
         return {
             "failed_documents": failed_documents,
-            "logging_collection": logging_collection,
+            "logging_collection": logging_dataset_id,
         }
 
     def insert_df(self, dataset_id, dataframe, *args, **kwargs):