+++ relevanceai/http_client.py
@@ -5,13 +5,11 @@ import json
 import os
 
 from doc_utils.doc_utils import DocUtils
-from relevanceai.dataset_api.dataset import Dataset
+from relevanceai.dataset_api.dataset import Dataset, Datasets
 
 from relevanceai.errors import APIError
 from relevanceai.api.client import BatchAPIClient
-from relevanceai.api.endpoints.services.cluster import ClusterClient
 from relevanceai.config import CONFIG
-from relevanceai.vector_tools.cluster import KMeans
 from relevanceai.vector_tools.plot_text_theme_model import build_and_plot_clusters
 
 
@@ -69,7 +67,8 @@ class Client(BatchAPIClient, DocUtils):
             )
         self.vector_tools = VectorTools(project, api_key)
 
-        self.Dataset = Dataset(self)
+        self.Dataset = Dataset(project=project, api_key=api_key)
+        self.Datasets = Datasets(project=project, api_key=api_key)
 
     # @property
     # def output_format(self):