+++ relevanceai/vector_tools/nearest_neighbours.py
@@ -21,6 +21,8 @@ class NearestNeighbours(_Base, DocUtils):
         callable_distance=None,
     ):
 
+        import scipy.spatial.distance as spatial_distance
+
         if callable_distance:
             sort_key = [
                 callable_distance(i, vector)