+++ tests/unit/test_clustering.py
@@ -37,6 +37,13 @@ def test_cluster_distribution(test_client, test_clustered_dataset):
     assert True
 
 
+def test_centroid_distances(test_client, test_clustered_dataset):
+    centroid_distances = test_client.vector_tools.cluster.centroid_distances(
+        test_clustered_dataset, "sample_1_vector_", "kmeans_10"
+    )
+    assert True
+
+
 @pytest.fixture
 def closest_to_centers(test_client, test_clustered_dataset):
     results = test_client.datasets.cluster.centroids.list_closest_to_center(